chore: refactored drug discovery tools
